+++ docs/metrics/getting-started/Program.cs
@@ -21,21 +21,20 @@ using OpenTelemetry.Metrics;
 public class Program
 {
     private static readonly Meter MyMeter = new Meter("MyCompany.MyProduct.MyLibrary", "1.0");
+    private static readonly Counter<long> MyFruitCounter = MyMeter.CreateCounter<long>("MyFruitCounter");
 
     public static void Main(string[] args)
     {
         using var meterProvider = Sdk.CreateMeterProviderBuilder()
-            .AddSource("MyCompany.MyProduct.MyLibrary")
+            .AddMeter("MyCompany.MyProduct.MyLibrary")
             .AddConsoleExporter()
             .Build();
 
-        var counter = MyMeter.CreateCounter<long>("MyFruitCounter");
-
-        counter.Add(1, new("name", "apple"), new("color", "red"));
-        counter.Add(2, new("name", "lemon"), new("color", "yellow"));
-        counter.Add(1, new("name", "lemon"), new("color", "yellow"));
-        counter.Add(2, new("name", "apple"), new("color", "green"));
-        counter.Add(5, new("name", "apple"), new("color", "red"));
-        counter.Add(4, new("name", "lemon"), new("color", "yellow"));
+        MyFruitCounter.Add(1, new("name", "apple"), new("color", "red"));
+        MyFruitCounter.Add(2, new("name", "lemon"), new("color", "yellow"));
+        MyFruitCounter.Add(1, new("name", "lemon"), new("color", "yellow"));
+        MyFruitCounter.Add(2, new("name", "apple"), new("color", "green"));
+        MyFruitCounter.Add(5, new("name", "apple"), new("color", "red"));
+        MyFruitCounter.Add(4, new("name", "lemon"), new("color", "yellow"));
     }
 }