+++ src/OpenTelemetry/Metrics/MetricStreamConfiguration.cs
@@ -21,16 +21,17 @@ namespace OpenTelemetry.Metrics
     {
 #pragma warning disable SA1602 // Enumeration items should be documented
         Default,
-        None,
+        Drop,
         Sum,
         LastValue,
         Histogram,
-        Drop = None,
 #pragma warning restore SA1602 // Enumeration items should be documented
     }
 
     public class MetricStreamConfiguration
     {
+        public static readonly MetricStreamConfiguration Drop = new DropConfiguration();
+
         public string Name { get; set; }
 
         public string Description { get; set; }