+++ test/OpenTelemetry.Tests.Stress.Metrics/Program.cs
@@ -15,7 +15,6 @@
 // </copyright>
 
 using System;
-using System.Diagnostics;
 using System.Diagnostics.Metrics;
 using System.Runtime.CompilerServices;
 using System.Threading;
@@ -40,7 +39,7 @@ public partial class Program
         }
 
         using var meterProvider = Sdk.CreateMeterProviderBuilder()
-            .AddSource("TestMeter")
+            .AddMeter("TestMeter")
             .Build();
         Stress();
     }