+++ src/OpenTelemetry.Api/CHANGELOG.md
@@ -5,6 +5,9 @@
 * Added `IDeferredMeterProviderBuilder`
   ([#2412](https://github.com/open-telemetry/opentelemetry-dotnet/pull/2412))
 
+* Breaking: Renamed `AddSource` to `AddMeter` on MeterProviderBuilder
+  to better reflect the intent of the method.
+
 ## 1.2.0-alpha4
 
 Released 2021-Sep-23